fix(changelog-generator): make matchOption() indicate failure
I am not sure exactly when this broke, but at some point we started
printing an error message for unrecognized options, but not actually
aborting. This commit restores the old behavior by propagating the
success status back to the caller via a boolean.

Noticed while extracting `matchOption()` into it's own file in the name
of decluttering, so this was originally going to be a "refactor" commit
but ended up being a "fix".