chore(amd-loader): add .yarnrc file
For consistency with other projects, this project is going to have the
same kind of Yarn-based release mechanism.

This one is "forward-thinking" in the same way that the issue template
is: we haven't moved the package into the `@liferay/` named scope on npm
yet, but we are going to do that real soon now, so this just gives us
one less thing to worry about when we do that.