chore(amd-loader): hoist devDependencies up to top level
Most of these are straight moves to keep this as minimal as possible,
but there is some really old stuff in there that I couldn't bring myself
to move, so there are some deviations:

-   babel-preset-es2015 is deprecated:

    https://www.npmjs.com/package/babel-preset-es2015

    so I switched to `@babel/preset-env`

-   publish-please is a replacement for `npm publish`, which we don't
    plan on using in this project anymore (ie. we'll just use `yarn
    version` to call `@liferay/js-publish`), so I dropped it.

-   I updated some versions to sync with stuff that we are using elsewhere
    already (eg. newer versions of `fs-extra` and `globby` etc) or stuff
    that seems to be pretty old (eg. `uglifyjs-webpack-plugin` was on a
    2018 version, so I bumped us to the latest, published in 2019)
    etc... I don't think there are any breaking changes for us in there,
    but if there are, I'll find out and fix those in subsequent commits.

combohandler hasn't been updated in 6 years, but it will take a bit of
time to find a replacement. My principal concern is that it will drag in
older cruft that is going to add noise to our `yarn audit` output.