test(amd-loader): prepare to run on GitHub actions
The top-level CI task in the monorepo is going to call `build` (via
`yarn workspaces run`) and then `test`, so we need a `build` task that
does all the things `ci` used to do.

In reality, the `ci` task is only used locally, so that you can check
locally the things that CI is going to check, but GitHub actions don't
ever call `ci`: they do a `build` plus `test` like I said above, and
there are separate actions that do `lint` and `format:check`.