chore(amd-loader): prepare to release under @liferay named scope
This is going to be one of the simpler projects to move, as far as I can
tell, because:

- I already named the directory "amd-loader" instead of
  "liferay-amd-loader" when I imported the history.
- The autolabeler is already set up to use the "amd-loader" label for
  that directory.
- The .yarnrc file is already set up to assume the new name.

But note that we must touch two other projects in this commit:

- The preset needs to know about the new name and version.
- The toolkit needs to know for the same reason.
- We update some docs.

One of the nice things about the monorepo is that we can make an atomic
commit like this that touches three projects in a single change. I
believe all this is safe because the changes to the loader are all safe
(ie. backportable), which means that in turn the changes to the toolkit
and the preset are safe (because they can be backported too).