chore(js-toolkit): make release process uniform
Do all of these the same way.

Note that we are not running `yarn ci` in preversion because we'll often
end up releasing more than one package in a batch, and we expect the
maintainer to see a green CI run (and probably a local one too) before
doing that. I expect that later on we'll have some Lerna-like automation
in place around this (but not Lerna itself, because its dependency
footprint is off the charts).