test(monorepo): consistently run local tests scoped to project/package
The idea is that when you run `yarn ci` from anywhere in the repo we
run all the tests in the repo to make sure that everything works well
together, but when you are inside a project or a package directory,
you can run `yarn test` to run only the tests for that package and
get a quicker feedback loop. Running `yarn test` at the top level
continues to run anything. Running it at an intermediate level, like
say, "projects/npm-tools/" will run it for all the packages nested below
that point.