feat(monorepo): add liferay-workspace-scripts
This new private package is to the monorepo what liferay-npm-scripts is
to the liferay-portal repo: that is, it presents a simplified interface
over the top of some common operations (build, format, lint, test etc)
that we can call from all over without having to pass complicated
parameter lists.

In terms of prior art, I did something similar in a monorepo over here
a while back:

    https://github.com/wincent/js/tree/master/packages/workspace-scripts

We talked about doing this in Clay, but never got around to it:

    https://github.com/liferay/clay/issues/1567

The solution taken in Clay is to use a Yeoman generator to produce its
40+ packages. This makes creation easy, but not maintenance, because if
you want to change what the scripts do, or what scripts there are, you
have to touch every single package.json file. Hence, the alternate
approach taken here.

Why not just use liferay-npm-scripts, then? Well, that is currently very
specifically tailored for use in liferay-portal (apart from some minor
use for formatting in this repo and in the liferay-learn one), and there
are some kinds of things we want to do in this repo that we'll never
want to do in liferay-portal, as far as I know (eg. publishing npm
packages). So, the plan then, is to continue investing in
liferay-npm-scripts and making it more generally useful, but to have a
specialized tool in this repo that calls liferay-npm-scripts whenever
appropriate (you can already see this because our `format` task calls
liferay-npm-scripts).

Test plan: ran a bunch of commands; I don't know if I managed to run
every command from every directory, but I did try to hit a good sample
of them... added some `console.log()` to things I can't really run yet
for real (like the `publish`) tasks to convince myself that they are
probably going to work.