refactor(js-insights): make package less hostile to testing
It's hard to test this package because it runs as soon as you require
it. So, create a `bin` entry that calls `main()` and runs it, instead of
just auto-running as soon as you require the `index.js`.

This means we'll be able to require `index.js` from a test and do
something useful with it.