test(js-insights): add "blinking light" tests
Again, these just show that the code does something at all, and serve as
a place where real tests could be added in the future. I just exposed
the `parse()` function because it was the smallest and most obvious
testable unit that I could find.

With this commit, every package in npm-tools has at least some tests.