test(js-toolkit): add "blinking light" tests npm-bridge-generator
This is the only package under the js-toolkit that doesn't have any
tests, so adding this just as a stub so that we can run them all in the
same way.

Speaking of "same", note that all the other projects in the monorepo
store there tests in "test/" to match the liferay-portal conventions,
but this project uses colocated "__tests__" files. We'll eventually want
to change all these to "test/" to make them globally consistent, but for
now I am adding this new test in "__tests__/" for local consistency.