test(js-toolkit): add package-local test scripts
For consistency with the rest of the repo. I believe that's every single
package done now, not counting the "maintenance" directory, which I have
no plans to touch at this time.

Test plan: Run `yarn run test` in every package directory. Note that in
the case of the "npm-bundler" tests, because Jest uses that a
substring-based filter, we slightly overtest in that case because we
also test the npm-bundler-preset-liferay-dev tests, which is an
innocuous enough thing:

   PASS  projects/npm-tools/packages/npm-bundler-preset-liferay-dev/test/index.js
   PASS  projects/js-toolkit/packages/npm-bundler/src/transform/text/operation/__tests__/replace.test.ts
   PASS  projects/js-toolkit/packages/npm-bundler/src/report/__tests__/index.test.ts