chore(npm-scripts)!: drop IE support from Babel config
BREAKING CHANGE: This commit takes our 7.3 compatibility matrix, and
drops IE from the list.

In order for this to be safely used in pre-7.4 versions, we will have to
add `.babelrc.js` config there that overwrites the settings added here.
(Well, that's the plan for now at least; if the overiding proves to be
too hard, I'll figure out another way.)

Note that Babel has always recommended being explicit anyway:

> Since one of the original goals of preset-env was to help users easily
> transition from using preset-latest, it behaves similarly when no
> targets are specified: preset-env will transform all ES2015-ES2020
> code to be ES5 compatible.
>
> We don't recommend using preset-env this way because it doesn't take advantage of its ability to target specific environments/versions.

And also:

> `preset-env`'s behavior is different than browserslist: it does _not_ use
> the `defaults` query when there are no targets are found in your Babel
> _or_ browserslist config(s). If you want to use the defaults query, you
> will need to explicitly pass it as a target:
>
> ```
> {
>   "presets": [["@babel/preset-env", {"targets": "defaults"}]]
> }
> ```

Both of those are from:

   https://babeljs.io/docs/en/babel-preset-env

Note that, even in 7.4, there will be places where we want to override,
such as analytics-client-js, because that is code that happens to
live in the liferay-portal repo but is not distributed with DXP
(rather, it is published separately via CDN, and we'll want to maintain
compatibility with IE there).

The compatibility matrix is currently at this link (not sure if it is
a permanent link or not, due to all the junk in the URL). In any case,
you can easily find it if you Google for ("Liferay DXP compatibility
matrix"):

    https://www.liferay.com/documents/10182/246659966/Liferay+DXP+7.3+Compatibility+Matrix.pdf/30fb6a1e-431d-7c84-e211-f2a0590013bf

It says:

-   Desktop Browser:
    -   Chrome 65
    -   Chrome (Latest)
    -   Edge (Latest)
    -   Firefox (Latest)
    -   Firefox ESR (Latest)
    -   Internet Explorer 11 (Limited)
    -   Safari 12
    -   Safari 13
-   Mobile Browser:
    -   Chrome (Latest)
    -   Safari (Latest)

Note that I am asking for "Safari >= 12" instead of "Safari 12, Safari
13", because Safari is already up to v14 and it seems absurd to ignore
it:

    https://en.wikipedia.org/wiki/Safari_version_history

Likewise, not from the Browserslist docs that mobile Safari translates
to "iOS", and mobile Chrome translates to "ChromeAndroid" (Chrome on iOS
is actually just a WebKit wrapper, so is covered by "iOS"):

    https://github.com/browserslist/browserslist

Test plan:

Run `npx browserslist 'Chrome 65, Firefox ESR, Safari >= 12, last 1
Chrome version, last 1 ChromeAndroid version, last 1 Edge version, last
1 Firefox version, last 1 iOS version'` and see:

    and_chr 85
    chrome 86
    chrome 65
    edge 86
    firefox 81
    firefox 78
    ios_saf 14
    safari 14
    safari 13.1
    safari 13
    safari 12.1
    safari 12

Deploy this in liferay-portal and check the generated output looks as
you would expect (eg. it has arrow functions in it).