feat(npm-scripts): automatically use legacy Babel config on older branches
The main choice I had here was between:

1. Making the "babel-legacy.json" file contain only overrides for the
   standard "babel.json" config. This would keep the overrides more
   minimal at the cost of adding a bit of complexity to the code (ie.
   instead of merge base config and user config, we'd have to merge base
   config, override config, and user config).
2. Making the "babel-legacy.json" file a full config in it's own right.
   This keeps the merging code simpler at the cost of having
   near-duplicate config in two files.

In the end, it probably doesn't matter much, so I went with "2". I
implemented "1" on top of this commit and you can see the diff here:

    https://gist.github.com/wincent/330ef35feb837f9a8e43b3cf97a27367

but it doesn't seem like much of an improvement to me. We can always
revisit it later if we change our minds.

The other choice here would be to have configs that look like:

- "babel.json"
- "babel-7.3.json"
- "babel-7.2.json"
- etc

But until we discover the need to differentiate at that level, I prefer
to keep the simpler "current" vs "legacy" distinction.