feat(npm-scripts): make JSP linting accept arrow functions on 7.4
We control this by setting the `ecmaVersion` setting based on the
current branch. 7.3 and friends get version 5, and 7.4 (or anywhere
else) get version 6.

Note that the tests need a bit of song and dance because the `lintJSP()`
function relies on private module state that is set up once when the
module is first evaluated (for speed, because we have to check hundreds
of JSP files in a full-repo lint run). So, we have to do
`jest.resetModules()` and make sure that we re-evaluate the `require()`
for every test.