feat(npm-scripts): add minification to build process for JS files
This implements the first part of the idea described here:

    https://docs.google.com/document/d/1JYSSSVFQlEwy1CRv1ZD6hfql8AzuXTS39qY5HMnvUe4/edit

Namely, if we're not in development mode, we pre-minify JS resources
before they are served. At the moment, I developed this just enough to
get some numbers. Here is what we see in `frontend-js-react-web`, which
is a good example of an OSGi module containing some "own JS" +
dependencies in `node_modules`.

    Summary:
      Files minified:    806
      Files with errors: 169
      Before size:       3561639
      After size:        878708
      Delta:             2682931
      Elapsed:           8.88

Sizes are just raw bytes, ignoring gzip, so the "real" size in a
production environment would be much less, obviously (but so will the
savings).

Lots of files with errors, but at a glance, most of those look legit,
because our bundler will take every file in the graph and AMD wrap it
even if it shouldn't be wrapped (due to config, or whatever), so we wind
up with illegal JS in the JAR that will never get used (like files with
`import` statements inside AMD factory definitions).

The approach is to minify all the JS as a last step in the build
process. This way we cover JS that came in from `yarn install` and the
bundler, JS that we wrote and built with Babel, and even JS that came
out of webpack. This way we don't have to figure out how to jam a
minifier into three places, and it also means that we might be able to
do some better minification (eg. by minifying pre-bundled source where
we are more likely to be able to detect useful amounts of dead code
etc). Note that this also helps "future proof" us a little bit for the
arrival of Bundler v3 or v4 or whatever it ends up being called, because
the approach can be applied there as well with few or no modifications.