feat(npm-scripts): minify JS in JSP
This is a WIP which will need some tweaking before it actually works,
because just as I foreshadowed here:

https://github.com/liferay/liferay-frontend-projects/pull/233#issuecomment-726058803

> that might be rather tricky though because the technique we use to run
> Prettier on JS inside JSP may not transfer across so well (because it
> is pretty whitespace-sensitive)

You can see this in the output of accounts-admin-web, were a file like

- src/main/resources/META-INF/resources/account_users_admin/select_account_entry.jsp
  (contents: https://gist.github.com/wincent/af4157dd55eecc3a862dc75cd560f572)

gets mangled to:

- build/node/packageRunBuild/resources/account_users_admin/select_account_entry.jsp
  (contents: https://gist.github.com/wincent/df3db72e83ac0d3e523b9bb7012417d3)

Here's the diff:

https://gist.github.com/wincent/fb57fdd77c3a694405487f38cf1b6a03

It actually doesn't look too shabby at all, but will need some tweaking
to get it totally right.

Notably not handled by this approach are sneaky JSP files that contain
JS outside of `<script>` (or `<aui:script>`) tags. For example:

- JSP: https://github.com/liferay/liferay-portal/blob/510836e72b25d47d8e7095d60506bfc34b104769/modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/available_languages.jsp
- Which is exposed sneakily here, as far as I can tell:
  https://github.com/liferay/liferay-portal/blob/510836e72b25d47d8e7095d60506bfc34b104769/portal-web/docroot/html/common/themes/top_js.jspf#L315