fix(npm-scripts): look for JSP files in classes/, not build/
This may end up being a throw-away commit — I am not sure whether this
is legit — or I will squash it into the parent before merging.

Yesterday when I tested this there were ".jsp" files in the "build"
folder, which is how I was able to produce this diff showing the
minified JS:

    https://gist.github.com/wincent/fb57fdd77c3a694405487f38cf1b6a03

Well, today after cleaning "build" and "classes", I only have JSP in
the latter. So I am hard-coding this modification in the most horrible
way, but it does work. The minified JS winds up in the JSP files that
you can find in the JAR; in the case of the module that I'm currently
testing, that's:

    build/tmp/jar/com.liferay.account.admin.web-2.0.0.jar

Doubts:

- Is this even remotely kosher?
- Should it be configurable?
- Will modifying files in "classes/" cause trouble (redundant builds
  stale caches) with Gradle or with "deploy fast" mode?