chore(npm-scripts): make minification more conservative
Based on testing across the liferay-portal repo, the property mangling
needs to be turned off because it does some IMO surprising things. For
example, in:

    https://github.com/liferay/liferay-portal/blob/c74cbb3487428c54ae2122c205af3c954e0e01a3/modules/apps/frontend-js/frontend-js-aui-web/src/main/resources/META-INF/resources/liferay/modules.js#L18-L21

This:

```
(function () {
   var LiferayAUI = Liferay.AUI;

   var COMBINE = LiferayAUI.getCombine();
```

gets mangled to:

```
!function(){var e = Liferay.AUI,a=e.i()
```

ie. global `Liferay.AUI` gets preserved (good), but a call to its
`getCombine()` method gets unfathomably rewritten to `i()`. No idea why
it would think it's legit to do that, but whatever, we can turn off that
kind of mangling with the following effects:

-   With unwanted mangling (ie. prior to this commit):

    Before size:       616616
    After size:        201246
    Delta:             415370

-   With `mangle: false` (ie. no mangling at all):

    Before size:       616616
    After size:        353013
    Delta:             263603

-   With `mangle: {properties: false}` (ie. this commit):

    Before size:       616616
    After size:        266414
    Delta:             350202

So we lose a bit of minification but not all of it (ie. instead of
trimming 405K, we trim 341K, which is not quite as bad as if we'd turned
off all mangling, which would mean trimming only 257K).

The other step in a conservative direction that I'm taking in this
commit is turning off some other renames that might cause problems when
minifying JSP. I haven't seen any evidence that this is needed yet, but
am erring on the side of caution.