fix(npm-scripts): try formatting JS in JSP inside build
Just testing the claim made in this Slack thread:

https://liferay.slack.com/archives/C04D8HMEF/p1606323974062700?thread_ts=1606233408.059000&cid=C04D8HMEF

that the reason why my earlier attempts to minify-in-place inside
"build/" may have seemed to not be working due to Gradle caching.

While the cache may cause problems, I am skeptical that this will work
because in my testing I found that `processResources` happens too late
(after `npm-scripts` runs), but in the interests of thoroughness I am
going to try this out and test it in a full `ant all` run anyway.