test(js-themes-toolkit): fix tests on non-Windows environment
Not sure why these used to pass on the old repo (which ran Travis) but
not on GitHub actions. Whatever the case, it is clear that in the
context of GitHub actions, Chalk thinks it is not running in an
interactive terminal, so it refuses to output color, thus breaking the
tests.

So, let's trick Chalk into doing what we want by setting the
`FORCE_COLOR` environment variable.

Note: plenty of Windows tests still failing (we also didn't use to run
those in Travis), so I will fix those or disable them in a separate
commit.