refactor(js-themes-toolikt): hoist lint and formatting config to repo root
Didn't have time to do this the other day when I did the import so
catching up now; this just gets rid of local ESLint config and delegates
the task of linting and formatting to the top-level, global scripts.

Note that I needed to expand the "files" patterns so that we correctly
identify files in "__tests__" directories as places where we can use
Jest globals like `it` and `describe`. We should standardize the pattern
here across the entire repo, but that will involve moving a *lot* of
stuff, so I'm leaving that for a PR another time. In fact, there are
quite a few trade-offs to consider so I've done a braindump for that in
this issue here:

    https://github.com/liferay/liferay-frontend-projects/issues/280