chore(js-themes-toolkit): make gulp-sass dependencies consistent
I just about tied my brain in knots making this small commit, but here
is what I checked before starting:

-   Places we currently use node-sass, as reported by `yarn why
    node-sass`:

    -   liferay-portal: node-sass 4.11.0
        -   via gulp-sass 3.2.1 (which depends on node-sass@^4.8.3)
            -   via liferay-theme-tasks
                -   via npm-scripts

    -   js-themes-toolkit 10.x series: node-sass 4.14.1
        -   via gulp-sass 3.2.1
            - via declared dependency on gulp-sass@^3.2.0

    -   js-themes-toolkit 9.x series: node-sass 4.13.1
        -   via gulp-sass 3.2.1
            - via declared dependency on gulp-sass@^3.2.0

    -   js-themes-toolkit 8.x series: node-sass 4.11.0
        -   via gulp-sass 3.2.1
            - via declared **dev** dependency on gulp-sass@^3.1.0
        -   via gulp-sass 3.2.0
            -   via liferay-theme-deps-7.0 and
                liferay-theme-deps-7.1
                -   via declared dependency on gulp-sass@3.2.0 (exact)

In terms of what versions of node-sass support what versions of NodeJS,
they have this lovely table:

    https://www.npmjs.com/package/node-sass

with a bunch of rows, but some interesting ones are:

-   Node 8 needs node-sass 4.5.3+, < 5.0
-   Node 10 needs node-sass 4.9+
-   Node 12 needs node-sass 4.12+

In other words, any version of node-sass >= 4.5.3 should work fine on
any version of Node but things get complicated when you hit node-sass
5.0; that version, which just came out a month ago, drops support for
all versions of Node except LTS versions (eg. 10, 12, 14) and the
current version (eg. 15).

Most of our customers probably use Node 10, but in case they don't, we
should not move to node-sass 5.0 at this time.

The js-themes-toolkit advertises itself as working with these versions
of DXP:

-   v8 = 6.2, 7.0, 7.1
-   v9 = 7.1 -> 7.2+ (gulp v3)
-   v10 = 7.1 -> 7.2+ (gulp v4)

And the version of Node that we use inside DXP was as follows (at least,
it was the last time I checked):

-   DXP 7.0.x: Node 6.6.0
-   DXP 7.1.x: Node 8.15.0 (note: even back here, async/await exists)
-   DXP 7.2.x: Node 10.15.1
-   DXP 7.3.x: Node 10.15.1
-   DXP master: 10.15.1

So, taking all that together, I think it is totally safe for us to
standardize on the latest gulp-sass (3.2.1) everywhere, and consequently
the latest node-sass in the 4.x series.

Steps taken in this commit:

1.  In 8.x, move gulp-sass from `devDependencies` to `dependencies` (ie.
    make it explicit instead of transitive, to match what we do on the other
    branches)
2.  In 8.x, update liferay-theme-deps-7.0 and liferay-theme-deps-7.1 to
    both depend on gulp-sass 3.2.1 instead of 3.2.0. I doubt we'll ever
    see another 3.x release of gulp-sass (3.2.1 was released 3 years
    ago), which means that this will keep us on node-sass 4.x for the
    foreseeable future.

As you can see, it produces a desirably small change in the lockfile,
including a move of node-sass up to 4.14.1. We have the option later on
of freshening lockfiles elsewhere (like in liferay-portal, and in the
9.x workspace) to bring them all into sync with 4.14.1.