test(js-themes-toolkit): run v8 and v9 tests in CI
Note that, just like we do for v2 of the js-toolkit, the `yarn ci` task
we're running here is only running tests, not formatting and linting,
because that is already covered elsewhere. The upside is that we don't
waste CI time performing redundant checks; the downside is that if
you're working in a local checkout you don't get the full battery of
checks by just running `yarn ci`.

This problem would be easily solved by making these special CI jobs use
a different task. That is probably ok, given that they already are
"special". I'll do that in the next commit for all three projects I
think.