refactor(monorepo): make maintenance project `yarn ci` run all checks
As explained in the previous commit, we had these maintenance projects
running just the "ci" task just like all the non-maintenance ones, for
consistency. But that meant that their "ci" tasks could not include
"format:check" and "lint" aspects because that would be redundant and
overlap with checks already done repo-wide at the top-level. Downside
of that is that `yarn ci` in a maintenance project in a local checkout
doesn't do all the checks like it would in a non-maintenance project.

Let's invert that trade-off and make `yarn ci` do the full thing locally
at the expense of having to spell out the individual steps in then
GitHub action config. Hopefully this isn't a dumb idea (one more thing
to keep in sync), but if it is, we can always revert it.