refactor(npm-scripts): use ProcessExitError instead of process.exit()
We want to print timing information to see how much time we're spending
in different parts of the build (or really, any task) in npm-scripts.

In order to do that, we basically need to be able to grab a timestamp at
the beginning, another at the end, and then do something with the delta.

However, there is one place in npm-scripts where we are "naughty" and
just bomb with `process.exit()` in order to signal a problem. This means
that the following construct (in pseudo-code) won't work:

    const start = now();

    try {
         run();
    } finally {
         const delta = now() - start;

         // If `process.exit()` was called,
         // we never get here and can't do
         // anything with `delta`.
    }

This means, for example, that you can't grab timing information for a
Prettier run that has formatting issues, even though that's a totally
legitimate use case (for example, we might want to time how long it
takes to fix formatting errors).

So, in this one naughty place, use an exception to signal our intent to
exit with a non-zero exit code. The `main()` process catches that and
calls `process.exit()` for us, but that's fine because by then our
`finally` block will already have had its chance to do what it wants.

Note that we don't just `throw` directly because we don't want a massive
stack trace to be spewed on the console just because of a routine
formatting error.