feat(js-themes-toolkit): expose r2 API
As discussed in the issue, we already have a copy of the liferay-r2 code
originally hosted here:

    https://github.com/liferay-labs/R2

which got inlined into the themes toolkit in:

    https://github.com/liferay/liferay-js-themes-toolkit/commit/a843516ffb18bbeb9b3c3b1c7de8a2622d233b12#diff-a016e05e9164926c610083bf8d96c2a4d473221ecda45c00522bdef0c5d97585

So, given that we needed access to r2's functionality in this PR:

    https://github.com/liferay/liferay-frontend-projects/pull/282

we can just leverage the existing copy of the code that we have in the
theme tasks (npm-scripts already depends on the theme tasks).

Note that I added a couple of simple tests to show that you can actually
grab and use the r2 API from the public interface of the package.

Closes: https://github.com/liferay/liferay-frontend-projects/issues/249