chore: hoist guidelines issue templates up to repo root
I think all of these pretty much make sense as top-level issue
templates, although we can tweak them any time if we see fit.

I considered making a "guidelines" template that covered all three, but
I'm not sure if the consolidation adds more than it takes away, so I
decided to leave it as-is.