feat(guidelines): mitigate problems with link checker rate limiting
This mitigates but does not fix:

    https://github.com/liferay/liferay-frontend-projects/issues/360

At this point we see a reliable pattern of npmjs servers rate-limiting
us with HTTP 429 "Too Many Requests" status codes:

    https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/429

Seeing as these links are not definitely broken (ie. they're just not
provably correct) we should treat them as soft warnings (even if no one
reads them) instead of hard errors.

We can still look at doing something fancy in the future to properly
deal with #360, but the present change is cheap and stops the bleeding.