feat(npm-scripts): don't generate main entry bridge if not needed
Before this commit we always generated the main entry bridge automatically, but we don't want
that for two reasons:

1. For the sake of consistency, it makes more sense to generate bridges only when told to do so,
whether it is setting `bridges` to `true` in `npmscripts.config.js` or defining a list of bridges in
array. That is, only generate the main entry bridge if `bridges` configuration is truthy.

2. Because we need to temporarily disable the automatic bridges in `liferay-portal` due to a bug
in webpack.