fix(npm-scripts): don't federate the whole React yet
This commit is like reverting to version 36.2.0 but without removing the code, so that we leave
`liferay-portal` build in a stable situation for further development.

The code being removed was necessary for LPS-124287 but given that it won't make it into portal
until a new webpack release is made, we are removing this code until then.

We are also refactoring the `CORE_REMOTES` and `CORE_SHARES` into a single
`FEDERATED_PACKAGES` object much better structured. See the comment above the variable
to get a more detailed explanation (I placed that in a comment to avoid future confusions if
we don't get to this commit; given that the objects structure is redundant for no apparent reason).