feat(npm-scripts): add more projects to federated build
Note that this configuration object is equivalent to the bundler's default preset
https://github.com/liferay/liferay-frontend-projects/blob/3fc3764838c85494e07a248f58bd5bba02044be4/projects/npm-tools/packages/npm-bundler-preset-liferay-dev/config.json#L22

We are including it inside `npm-scripts` for now, but we will need to extract it somewhere in the
future (if not for us, for customers, to configure their builds). The way to do it is yet to be defined.