refactor(senna): switch to Yarn
This is the result of running `yarn` and deleting the package-lock.json
file. Note that the build is horribly broken both before and after this
change due to the usual node-sass shenanigans (which we partake in by
virtue of using gulp-metal, which pulls in gulp-sass):

    Error: Node Sass does not yet support your current environment: OS X
    64-bit with Unsupported runtime (64)

I think I'll be able to fix this by pinning us to a later release of
node-sass, but that in itself might be tricky:

- Senna depends on `gulp-metal@^1.0.0`; the latest release of gulp-metal
  is 2.2.2 (from mid-2017), but Senna is still using 1.12.2 (from around
  the same time; I'm not sure why we didn't update back then).
- gulp-metal 1.12.2 depends on `gulp-sass@^2.0.1`, which resolves to
  2.3.2 (from mid-2016), and which in turn depends on `node-sass@^3.4.2`
  (resolves to v3.13.1, from late-2016).
- Additionally, 2.2.2 of gulp-metal still depends on a too-old version
  of gulp-sass (`^3.1.0`), so updating to the last gulp-metal wouldn't
  help us here. If we update to `gulp-metal@2.2.2`, we get
  `gulp-sass@3.2.1` (early-2018), which depends on `node-sass@^4.8.3`,
  but somehow we still get `node-sass@4.5.3`, which will issue the same
  error seen above. (No idea why that might be, but I am 100% it is
  happening, having tested it twice now after a `git clean -fdx .`)

Given that this is a dev-only dep, it would be convenient to sketchily
force it to use a v4 release of node-sass via Yarn resolutions, but
unfortunately resolutions aren't powerful enough to override a hard
limit within a dependency range, only to select a version within such a
range (ie. we can't turn a `^3.4.2` dep into a `^4.14.1` one), so we'll
have to see if there is a way around this without having to do anything
unseemly (like cutting a release of gulp-metal, which is hard because
most of the owners on the npm package have left the company or are on
vacation, or inlining that project in here etc...)